Split Flows configuration in a logic and confiugration file
