+++ client/signup/config/flows.js
@@ -8,297 +8,74 @@ import i18n from 'i18n-calypso';
  * Internal dependencies
  */
 import { abtest } from 'lib/abtest';
-import config from 'config';
 import stepConfig from './steps';
 import userFactory from 'lib/user';
+import flowsData from './flows-configuration';
 
 const user = userFactory();
 
-function getCheckoutUrl( dependencies ) {
-	return '/checkout/' + dependencies.siteSlug;
-}
+class Flows {
+	defaultFlowName = 'main';
+	resumingFlow = false;
 
-function dependenciesContainCartItem( dependencies ) {
-	return dependencies.cartItem || dependencies.domainItem || dependencies.themeItem;
-}
-
-function getSiteDestination( dependencies ) {
-	if ( dependenciesContainCartItem( dependencies ) ) {
-		return getCheckoutUrl( dependencies );
+	constructor() {
+		if ( ! ( this instanceof Flows ) ) {
+			return new Flows();
+		}
 	}
 
-	let protocol = 'https';
+	filterFlowName( flowName ) {
+		const defaultFlows = [ 'main', 'website' ];
+
+		/**
+		 * Only run the User First Signup for logged out users.
+		 */
+		if ( ! user.get() ) {
+			if ( includes( defaultFlows, flowName ) && abtest( 'userFirstSignup' ) === 'userFirst' ) {
+				return 'user-first';
+			}
+		}
+
+		return flowName;
+	}
+
+	filterDesignTypeInFlow( flow ) {
+		if ( ! flow ) {
+			return;
+		}
+
+		if ( ! includes( flow.steps, 'design-type' ) || 'designTypeWithStore' !== abtest( 'signupStore' ) ) {
+			return flow;
+		}
+
+		return assign( {}, flow, {
+			steps: flow.steps.map( stepName => stepName === 'design-type' ? 'design-type-with-store' : stepName )
+		} );
+	}
 
 	/**
-	 * It is possible that non-wordpress.com sites are not HTTPS ready.
+	 * Filters the Flow destination if needed.
 	 *
-	 * Redirect them
+	 * Usually empty, but it can be used to filter the destination in order to do AB tests or override it.
+	 *
+	 * @param {String} destination The target destination
+	 * @param {Object} dependencies Signup's dependencies
+	 * @param {String} flowName The name of the flow
+	 * @returns {String} The flow where the user will be redirected to.
 	 */
-	if ( ! dependencies.siteSlug.match( /wordpress\.[a-z]+$/i ) ) {
-		protocol = 'http';
+	filterDestination( destination, dependencies, flowName ) {
+		return destination;
 	}
 
-	return protocol + '://' + dependencies.siteSlug;
-}
-
-function getPostsDestination( dependencies ) {
-	if ( dependenciesContainCartItem( dependencies ) ) {
-		return getCheckoutUrl( dependencies );
-	}
-
-	return '/posts/' + dependencies.siteSlug;
-}
-
-const flows = {
-	account: {
-		steps: [ 'user' ],
-		destination: '/',
-		description: 'Create an account without a blog.',
-		lastModified: '2015-07-07'
-	},
-
-	business: {
-		steps: [ 'design-type', 'themes', 'domains', 'user' ],
-		destination: function( dependencies ) {
-			return '/plans/select/business/' + dependencies.siteSlug;
-		},
-		description: 'Create an account and a blog and then add the business plan to the users cart.',
-		lastModified: '2016-06-02',
-		meta: {
-			skipBundlingPlan: true
+	removeUserStepFromFlow( flow ) {
+		if ( ! flow ) {
+			return;
 		}
-	},
 
-	premium: {
-		steps: [ 'design-type', 'themes', 'domains', 'user' ],
-		destination: function( dependencies ) {
-			return '/plans/select/premium/' + dependencies.siteSlug;
-		},
-		description: 'Create an account and a blog and then add the premium plan to the users cart.',
-		lastModified: '2016-06-02',
-		meta: {
-			skipBundlingPlan: true
-		}
-	},
-
-	free: {
-		steps: [ 'design-type', 'themes', 'domains', 'user' ],
-		destination: getSiteDestination,
-		description: 'Create an account and a blog and default to the free plan.',
-		lastModified: '2016-06-02'
-	},
-
-	'with-theme': {
-		steps: [ 'domains-theme-preselected', 'plans', 'user' ],
-		destination: getSiteDestination,
-		description: 'Preselect a theme to activate/buy from an external source',
-		lastModified: '2016-01-27'
-	},
-
-	subdomain: {
-		steps: [ 'design-type', 'themes', 'domains', 'plans', 'user' ],
-		destination: getSiteDestination,
-		description: 'Provide a vertical for subdomains',
-		lastModified: '2016-10-31'
-	},
-
-	main: {
-		steps: [ 'design-type', 'themes', 'domains', 'plans', 'user' ],
-		destination: getSiteDestination,
-		description: 'The current best performing flow in AB tests',
-		lastModified: '2016-05-23'
-	},
-
-	surveystep: {
-		steps: [ 'survey', 'design-type', 'themes', 'domains', 'plans', 'user' ],
-		destination: getSiteDestination,
-		description: 'The current best performing flow in AB tests',
-		lastModified: '2016-05-23'
-	},
-
-	website: {
-		steps: [ 'design-type', 'themes', 'domains', 'plans', 'user' ],
-		destination: getSiteDestination,
-		description: 'This flow was originally used for the users who clicked "Create Website" ' +
-			'on the two-button homepage. It is now linked to from the default homepage CTA as ' +
-			'the main flow was slightly behind given translations.',
-		lastModified: '2016-05-23'
-	},
-
-	blog: {
-		steps: [ 'design-type', 'themes', 'domains', 'plans', 'user' ],
-		destination: getSiteDestination,
-		description: 'This flow was originally used for the users who clicked "Create Blog" on ' +
-			'the two-button homepage. It is now used from blog-specific landing pages so that ' +
-			'verbiage in survey steps refers to "blog" instead of "website".',
-		lastModified: '2016-05-23'
-	},
-
-	personal: {
-		steps: [ 'design-type', 'themes', 'domains', 'user' ],
-		destination: function( dependencies ) {
-			return '/plans/select/personal/' + dependencies.siteSlug;
-		},
-		description: 'Create an account and a blog and then add the personal plan to the users cart.',
-		lastModified: '2016-01-21'
-	},
-
-	'test-site': {
-		steps: config( 'env' ) === 'development' ? [ 'site', 'user' ] : [ 'user' ],
-		destination: '/',
-		description: 'This flow is used to test the site step.',
-		lastModified: '2015-09-22'
-	},
-
-	'delta-discover': {
-		steps: [ 'user' ],
-		destination: '/',
-		description: 'A copy of the `account` flow for the Delta email campaigns. Half of users who ' +
-			'go through this flow receive a reader-specific drip email series.',
-		lastModified: '2016-05-03'
-	},
-
-	'delta-blog': {
-		steps: [ 'design-type', 'themes', 'domains', 'plans', 'user' ],
-		destination: getSiteDestination,
-		description: 'A copy of the `blog` flow for the Delta email campaigns. Half of users who go ' +
-			'through this flow receive a blogging-specific drip email series.',
-		lastModified: '2016-03-09'
-	},
-
-	'delta-site': {
-		steps: [ 'design-type', 'themes', 'domains', 'plans', 'user' ],
-		destination: getSiteDestination,
-		description: 'A copy of the `website` flow for the Delta email campaigns. Half of users who go ' +
-			'through this flow receive a website-specific drip email series.',
-		lastModified: '2016-03-09'
-	},
-
-	desktop: {
-		steps: [ 'design-type', 'themes', 'domains', 'plans', 'user' ],
-		destination: getPostsDestination,
-		description: 'Signup flow for desktop app',
-		lastModified: '2016-05-30'
-	},
-
-	developer: {
-		steps: [ 'themes', 'site', 'user' ],
-		destination: '/devdocs/welcome',
-		description: 'Signup flow for developers in developer environment',
-		lastModified: '2015-11-23'
-	},
-
-	pressable: {
-		steps: [ 'design-type-with-store', 'themes', 'domains', 'plans', 'user' ],
-		destination: getSiteDestination,
-		description: 'Signup flow for testing the pressable-store step',
-		lastModified: '2016-06-27'
-	},
-
-	jetpack: {
-		steps: [ 'jetpack-user' ],
-		destination: '/'
-	},
-
-	'get-dot-blog': {
-		steps: [ 'get-dot-blog-themes', 'get-dot-blog-plans' ],
-		destination: getSiteDestination,
-		description: 'Used by `get.blog` users that connect their site to WordPress.com',
-		lastModified: '2016-11-14'
-	},
-
-	'user-first': {
-		steps: [ 'user', 'design-type', 'themes', 'domains', 'plans' ],
-		destination: getSiteDestination,
-		description: 'User-first signup flow',
-		lastModified: '2016-01-18',
-	},
-};
-
-if ( config.isEnabled( 'signup/domain-first-flow' ) ) {
-	flows[ 'domain-first' ] = {
-		steps: [ 'site-or-domain', 'themes', 'plans', 'user' ],
-		destination: getSiteDestination,
-		description: 'An experimental approach for WordPress.com/domains',
-		lastModified: '2017-01-16'
-	};
-
-	flows[ 'site-selected' ] = {
-		steps: [ 'themes-site-selected', 'plans-site-selected' ],
-		destination: getSiteDestination,
-		providesDependenciesInQuery: [ 'siteSlug', 'siteId' ],
-		description: 'A flow to test updating an existing site with `Signup`',
-		lastModified: '2017-01-19'
-	};
-}
-
-if ( config.isEnabled( 'signup/social' ) ) {
-	flows.social = {
-		steps: [ 'user-social' ],
-		destination: '/',
-		description: 'Create an account without a blog with social signup enabled.',
-		lastModified: '2017-03-16'
-	};
-}
-
-if ( config( 'env' ) === 'development' ) {
-	flows[ 'test-plans' ] = {
-		steps: [ 'site', 'plans', 'user' ],
-		destination: getSiteDestination,
-		description: 'This flow is used to test plans choice in signup',
-		lastModified: '2016-06-30'
-	};
-}
-
-function removeUserStepFromFlow( flow ) {
-	if ( ! flow ) {
-		return;
+		return assign( {}, flow, {
+			steps: reject( flow.steps, stepName => stepConfig[ stepName ].providesToken )
+		} );
 	}
-
-	return assign( {}, flow, {
-		steps: reject( flow.steps, stepName => stepConfig[ stepName ].providesToken )
-	} );
-}
-
-function filterDesignTypeInFlow( flow ) {
-	if ( ! flow ) {
-		return;
-	}
-
-	if ( ! includes( flow.steps, 'design-type' ) ) {
-		return flow;
-	}
-
-	return assign( {}, flow, {
-		steps: flow.steps.map( stepName => stepName === 'design-type' ? 'design-type-with-store' : stepName )
-	} );
-}
-
-function filterFlowName( flowName ) {
-	const defaultFlows = [ 'main', 'website' ];
-
-	/**
-	 * Only run the User First Signup for logged out users.
-	 */
-	if ( ! user.get() ) {
-		if ( includes( defaultFlows, flowName ) && abtest( 'userFirstSignup' ) === 'userFirst' ) {
-			return 'user-first';
-		}
-	}
-
-	return flowName;
-}
-
-function filterDestination( destination ) {
-	return destination;
-}
-
-const Flows = {
-	filterFlowName,
-	filterDestination,
-
-	defaultFlowName: 'main',
-	resumingFlow: false,
 
 	/**
 	 * Get certain flow from the flows configuration.
@@ -313,8 +90,8 @@ const Flows = {
 	 * @param {String} currentStepName The current step. See description above
 	 * @returns {Object} A flow object
 	 */
-	getFlow( flowName, currentStepName = '' ) {
-		let flow = Flows.getFlows()[ flowName ];
+	getFlow = ( flowName, currentStepName = '' ) => {
+		let flow = this.getFlows()[ flowName ];
 
 		// if the flow couldn't be found, return early
 		if ( ! flow ) {
@@ -322,22 +99,22 @@ const Flows = {
 		}
 
 		if ( user.get() ) {
-			flow = removeUserStepFromFlow( flow );
+			flow = this.removeUserStepFromFlow( flow );
 		}
 
 		// Show design type with store option only to new users with EN locale.
 		if ( ! user.get() && 'en' === i18n.getLocaleSlug() ) {
-			flow = filterDesignTypeInFlow( flow );
+			flow = this.filterDesignTypeInFlow( flow );
 		}
 
-		Flows.preloadABTestVariationsForStep( flowName, currentStepName );
+		this.preloadABTestVariationsForStep( flowName, currentStepName );
 
-		return Flows.getABTestFilteredFlow( flowName, flow );
-	},
+		return this.getABTestFilteredFlow( flowName, flow );
+	};
 
 	getFlows() {
-		return flows;
-	},
+		return flowsData;
+	}
 
 	/**
 	 * Preload AB Test variations after a certain step has been completed.
@@ -350,7 +127,7 @@ const Flows = {
 	 * @param {String} flowName The current flow
 	 * @param {String} stepName The step that is being completed right now
 	 */
-	preloadABTestVariationsForStep( flowName, stepName ) {
+	preloadABTestVariationsForStep = ( flowName, stepName ) => {
 		/**
 		 * In cases where the flow is being resumed, the flow must not be changed from what the user
 		 * has seen before.
@@ -358,7 +135,7 @@ const Flows = {
 		 * E.g. A user is resuming signup from before the test was added. There is no need
 		 * to add a step somewhere back in the line.
 		 */
-		if ( Flows.resumingFlow ) {
+		if ( this.resumingFlow ) {
 			return;
 		}
 
@@ -372,10 +149,10 @@ const Flows = {
 		 */
 		if ( 'main' === flowName ) {
 			if ( '' === stepName ) {
-				abtest( 'signupSurveyStep' );
+				abtest( 'siteTitleStep' );
 			}
 		}
-	},
+	};
 
 	/**
 	 * Return a flow that is modified according to the ABTest rules.
@@ -389,7 +166,7 @@ const Flows = {
 	 *
 	 * @return {Object} A filtered flow object
 	 */
-	getABTestFilteredFlow( flowName, flow ) {
+	getABTestFilteredFlow = ( flowName, flow ) => {
 		// Only do this on the main flow
 		if ( 'main' === flowName ) {
 			if ( abtest( 'signupSurveyStep' ) === 'showSurveyStep' ) {
@@ -398,7 +175,7 @@ const Flows = {
 		}
 
 		return flow;
-	},
+	};
 
 	/**
 	 * Insert a step into the flow.
@@ -431,7 +208,7 @@ const Flows = {
 		}
 
 		return flow;
-	},
+	}
 
 	removeStepFromFlow( stepName, flow ) {
 		return {
@@ -441,6 +218,6 @@ const Flows = {
 			} )
 		};
 	}
-};
+}
 
-export default Flows;
+export default new Flows();
